fix uie eval
